requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.1
pymongo==4.6.1
bcrypt==4.1.2
mysql-connector-python==8.3.0
google-generativeai==0.5.2
gunicorn==21.2.0
